Web_App/requirements.txt
Flask
flask-restful
google-cloud-datastore
google-auth
werkzeug
google-cloud-secret-manager
requests
pytz